Revert to single file
