--- lib/utf8_sanitizer.rb
@@ -1,7 +1,0 @@
-require 'utf8_sanitizer/error'
-
-require 'utf8_sanitizer/replace'
-require 'utf8_sanitizer/exception'
-
-module UTF8Sanitizer
-end
