--- lib/utf8_sanitizer/error.rb
@@ -1,4 +1,0 @@
-module UTF8Sanitizer
-  class Error < StandardError; end
-end
-
